+++ src/components/conditional.js
@@ -6,23 +6,30 @@
   jsx: (
     <div className={children.length === 0 ? classes.empty : undefined}>
       {(() => {
+        const { left, right, compare, visible: initVisibility } = options;
         const { useText, env } = B;
         const isDev = env === 'dev';
         const isPristine = isDev && children.length === 0;
+        const mounted = useRef(false);
+        const [leftValue, setLeftValue] = useState(useText(left));
+        const [rightValue, setRightValue] = useState(useText(right));
+        const [visible, setVisible] = useState(initVisibility);
 
         const evalCondition = () => {
-          const left = useText(options.left);
-          const right = useText(options.right);
-          const leftAsNumber = parseFloat(left);
-          const rightAsNumber = parseFloat(right);
+          const leftAsNumber = parseFloat(leftValue);
+          const rightAsNumber = parseFloat(rightValue);
 
-          switch (options.compare) {
+          if (!initVisibility && leftValue === '' && rightValue === '') {
+            return false;
+          }
+
+          switch (compare) {
             case 'neq':
-              return left !== right;
+              return leftValue !== rightValue;
             case 'contains':
-              return left.indexOf(right) > -1;
+              return leftValue.indexOf(rightValue) > -1;
             case 'notcontains':
-              return left.indexOf(right) < 0;
+              return leftValue.indexOf(rightValue) < 0;
             case 'gt':
               return leftAsNumber > rightAsNumber;
             case 'lt':
@@ -32,17 +39,14 @@
             case 'lteq':
               return leftAsNumber <= rightAsNumber;
             default:
-              return left === right;
+              return leftValue === rightValue;
           }
         };
+
         const checkCondition = evalCondition();
-        const initialVisibility = options.visible
-          ? checkCondition
-          : !checkCondition;
-        const [visible, setVisible] = useState(false);
 
         useEffect(() => {
-          setVisible(initialVisibility);
+          setVisible(checkCondition);
         }, [checkCondition]);
 
         useEffect(() => {
@@ -51,8 +55,17 @@
           } else {
             B.triggerEvent('isFalse', false);
           }
-          B.triggerEvent('onChange', visible);
+          if (mounted.current) {
+            B.triggerEvent('onChange', visible);
+          }
         }, [visible]);
+
+        useEffect(() => {
+          mounted.current = true;
+          return () => {
+            mounted.current = false;
+          };
+        }, []);
 
         B.defineFunction('Hide', () => setVisible(false));
         B.defineFunction('Show', () => setVisible(true));
@@ -60,6 +73,16 @@
         B.defineFunction('Set Visibility', value => {
           if (typeof value === 'boolean') setVisible(value);
         });
+
+        const getValue = evt => {
+          const value = (evt && evt.target && evt.target.value) || evt;
+          return `${value}`;
+        };
+
+        B.defineFunction('Set Left Value', evt => setLeftValue(getValue(evt)));
+        B.defineFunction('Set Right Value', evt =>
+          setRightValue(getValue(evt)),
+        );
 
         if (!isDev && !visible) return null;
         return isPristine ? 'Conditional' : children;