+++ src/components/textField.js
@@ -57,15 +57,22 @@
       id: customModelAttributeId,
       label = [],
       value: defaultValue = [],
+      required: defaultRequired = false,
     } = customModelAttributeObj;
-    const [currentValue, setCurrentValue] = useState(useText(defaultValue));
+    const [currentValue, setCurrentValue] = useState(
+      useText(defaultValue, { rawValue: true }),
+    );
     const labelText = useText(label);
     const customModelAttribute = getCustomModelAttribute(
       customModelAttributeId,
     );
 
-    const { name: customModelAttributeName, validations: { required } = {} } =
-      customModelAttribute || {};
+    const {
+      name: customModelAttributeName,
+      validations: { required: attributeRequired } = {},
+    } = customModelAttribute || {};
+
+    const required = customModelAttribute ? attributeRequired : defaultRequired;
     const nameAttributeValue = useText(nameAttribute);
 
     const validPattern = pattern || null;
@@ -167,9 +174,16 @@
       handleValidation(validity);
     };
 
+    useEffect(() => {
+      B.triggerEvent('onChange', currentValue);
+    }, [currentValue]);
+
     B.defineFunction('Clear', () => setCurrentValue(''));
     B.defineFunction('Enable', () => setIsDisabled(false));
     B.defineFunction('Disable', () => setIsDisabled(true));
+    B.defineFunction('Reset', () =>
+      setCurrentValue(useText(defaultValue, { rawValue: true })),
+    );
 
     const handleClickShowPassword = () => {
       togglePassword(!showPassword);