+++ src/interactions/redirect.ts
@@ -1,10 +1,25 @@
 interface Page {
   name: string;
+  resolvePath: (input: Record<string, string>) => string;
   url: string;
 }
 
-function redirect({ redirectTo }: { redirectTo: Page }): void {
-  const { url } = redirectTo;
+function redirect({
+  event,
+  redirectTo,
+}: {
+  event: Event;
+  redirectTo: Page;
+}): void {
+  let { url } = redirectTo;
+
+  if (
+    typeof event === 'object' &&
+    event !== null &&
+    typeof (event as any).id === 'number'
+  ) {
+    url = redirectTo.resolvePath(event as any);
+  }
 
   window.location.href = url;
 }